Move all test to new folder
