+++ tests/test_configuration.py
