+++ tests/test_potential.py
