+++ tests/test_selection.py
