+++ tests/test_wham.py
@@ -2,7 +2,7 @@ import os
 import numpy as np
 from mlptrain.sampling.bias import Bias
 from mlptrain.sampling.umbrella import _Window, UmbrellaSampling
-from .utils import work_in_zipped_dir
+from .data.utils import work_in_zipped_dir
 
 here = os.path.dirname(os.path.abspath(__file__))
 
@@ -40,7 +40,7 @@ def _initialised_us() -> UmbrellaSampling:
     return us
 
 
-@work_in_zipped_dir(os.path.join(here, 'data.zip'))
+@work_in_zipped_dir(os.path.join(here, 'data/data.zip'))
 def test_wham_is_close_to_ref():
     us = _initialised_us()
     zetas, free_energies = us.wham(n_bins=499)
@@ -65,7 +65,7 @@ def test_wham_is_close_to_ref():
         assert np.isclose(free_energy, ref_free_energies[close_idx], atol=0.02)
 
 
-@work_in_zipped_dir(os.path.join(here, 'data.zip'))
+@work_in_zipped_dir(os.path.join(here, 'data/data.zip'))
 def test_wham_is_somewhat_independent_of_nbins():
     us = _initialised_us()
     _, free_energies_500bins = us.wham(n_bins=500)
